requirementsFlask.txt
# requirementFlask.txt

Pillow 
Flask